+++ src/components/ChatMessage/index.css
@@ -8,33 +8,8 @@
     border-radius: 10px;
 }
 
-.message-box {
-    background-color: #f5f5f5;    
-    height: 50px;
-    margin: 10px;
-    display: flex;
-
-}
-
 .white-border {
     border-color: white;
-}
-
-.transparent-background {
-    background-color: transparent;
-}
-
-.input-box {
-    width: 46vw;
-    margin: 10px;
-    border: none;
-}
-
-.send-button {
-    margin: 5px;
-    background: none;
-    border: none;
-    box-shadow: none;
 }
 
 .message {
