+++ src/components/ChatWindow/index.js
@@ -1,35 +1,52 @@
-import React, { useState } from 'react';
+import { useState, useEffect } from 'react';
 import { socket } from '../../socket';
-import img from "../../assets/paper.png";
+import { MessageBox } from '../MessageBox';
+import { MessageWindow } from '../MessageWindow/Index';
+import { ConnectionStatus } from "../ConnectionStatus";
 import "./index.css";
 
-function sendJoinEvent(username, setIsConnected) {
-  const userData = { 'username': username }
-  socket.emit('join', userData);
-  setIsConnected(true)
-}
 
-export function MyForm() {
-  const [value, setValue] = useState('');
-  const [username, setUsername] = useState('');
-  const [isConnected, setIsConnected] = useState(socket.connected);
+// import { ChatMessage } from '../ChatMessage';
+// import { ConnectionState } from './components/ConnectionState';
+// import { ConnectionManager } from '../ConnectionManager';
 
-  function onSubmit(event) {
-    event.preventDefault();
-    isConnected ?
-      socket.emit('message', { user: username, 'message': value })
-      : sendJoinEvent(username, setIsConnected)
-      setValue('')
-  }
+export function ChatWindow() {
+  const [messageResponse, setMessageResponse] = useState([]);
+  const [connectionState, setConnectionState] = useState(false);
+
+  useEffect(() => {
+    function onMessageResponse(value) {
+      setMessageResponse(previous => [...previous, value]);
+    }
+
+    function onConnect() {
+      setConnectionState(true);
+    }
+
+    function onDisconnect() {
+      setConnectionState(false);
+    }
+
+    socket.on('disconnect', onDisconnect);
+    socket.on('connect', onConnect);
+    socket.on('messageResponse', onMessageResponse);
+
+    return () => {
+      socket.off('connect', onConnect);
+      socket.off('disconnect', onDisconnect);
+      socket.off('messageResponse', onMessageResponse);
+    };
+  }, []);
 
   return (
-    <form onSubmit={onSubmit}>
-      <div className='message-box'>
-      {isConnected ?
-        <input className= 'input-box curved-radius' value={value} onChange={e => setValue(e.target.value)} placeholder='Write a message...' />
-        : <input className= 'input-box curved-radius white-border'  value={username} onChange={(event) => setUsername(event.target.value)} placeholder='Type in a username here...' />}
-      <button type="submit" className='send-button curved-radius white-border transparent-background'> <img src={img} alt="send" width={20}/> </button>
+    <div className="App">
+      <div className='chat-window'>
+        <div className='messages-window'>
+          <MessageWindow messages={messageResponse} />
+        </div>
+        <MessageBox />
+        <ConnectionStatus connectionState={connectionState}/>
       </div>
-    </form>
+    </div>
   );
 }
