+++ src/components/MessageBox/index.css
@@ -1,0 +1,23 @@
+.message-box {
+    background-color: #f5f5f5;    
+    height: 50px;
+    margin: 10px;
+    display: flex;
+
+}
+
+.input-box {
+    width: 46vw;
+    margin: 10px;
+    border: none;
+}
+
+.send-button {
+    margin: 5px;
+    background: none;
+    border: none;
+    box-shadow: none;
+}
+.transparent-background {
+    background-color: transparent;
+}
